+++ rusoto/credential/src/instance_metadata.rs
@@ -1,9 +1,11 @@
 //! The Credentials Provider for an AWS Resource's IAM Role.
 
-use reqwest;
-use std::io::Read;
+use hyper::Uri;
+use std::time::Duration as StdDuration;
+use tokio_core::reactor::Core;
 
-use {AwsCredentials, CredentialsError, ProvideAwsCredentials, parse_credentials_from_aws_service};
+use {AwsCredentials, CredentialsError, ProvideAwsCredentials, ProvideTimeoutableAwsCredentials,
+     make_request, parse_credentials_from_aws_service};
 
 const AWS_CREDENTIALS_PROVIDER_IP: &'static str = "169.254.169.254";
 const AWS_CREDENTIALS_PROVIDER_PATH: &'static str = "latest/meta-data/iam/security-credentials";
@@ -14,36 +16,30 @@ pub struct InstanceMetadataProvider;
 
 impl ProvideAwsCredentials for InstanceMetadataProvider {
     fn credentials(&self) -> Result<AwsCredentials, CredentialsError> {
-        get_credentials_from_role()
+        get_credentials_from_role(None)
+    }
+}
+
+impl ProvideTimeoutableAwsCredentials for InstanceMetadataProvider {
+    fn credentials_with_timeout(
+        &self,
+        timeout: StdDuration,
+    ) -> Result<AwsCredentials, CredentialsError> {
+        get_credentials_from_role(Some(timeout))
     }
 }
 
 /// Gets the role name to get credentials for using the IAM Metadata Service (169.254.169.254).
-fn get_role_name() -> Result<String, CredentialsError> {
+fn get_role_name(timeout: Option<StdDuration>) -> Result<String, CredentialsError> {
     let role_name_address = format!(
-        "http://{}/{}",
+        "http://{}/{}/",
         AWS_CREDENTIALS_PROVIDER_IP,
         AWS_CREDENTIALS_PROVIDER_PATH
     );
-    let response = reqwest::get(&role_name_address);
-
-    match response {
-        Ok(mut resp) => {
-            if resp.status().is_success() {
-                let mut role_name = String::new();
-                if resp.read_to_string(&mut role_name).is_err() {
-                    return Err(CredentialsError::new(
-                        "Didn't get a parsable role name from metadata service",
-                    ));
-                }
-                Ok(role_name)
-            } else {
-                return Err(CredentialsError::new(format!(
-                    "Couldn't connect to credentials provider: {}",
-                    resp.status()
-                )));
-            }
-        }
+    let core = try!(Core::new());
+    let uri = try!(role_name_address.parse::<Uri>());
+    match make_request(core, uri, timeout) {
+        Ok(resp) => Ok(resp),
         Err(err) => {
             return Err(CredentialsError::new(
                 format!("Couldn't connect to credentials provider: {}", err),
@@ -53,8 +49,10 @@ fn get_role_name() -> Result<String, CredentialsError> {
 }
 
 /// Gets the credentials for an EC2 Instances IAM Role.
-fn get_credentials_from_role() -> Result<AwsCredentials, CredentialsError> {
-    let role_name = try!(get_role_name());
+fn get_credentials_from_role(
+    timeout: Option<StdDuration>,
+) -> Result<AwsCredentials, CredentialsError> {
+    let role_name = try!(get_role_name(timeout.clone()));
     let credentials_provider_url = format!(
         "http://{}/{}/{}",
         AWS_CREDENTIALS_PROVIDER_IP,
@@ -62,26 +60,10 @@ fn get_credentials_from_role() -> Result<AwsCredentials, CredentialsError> {
         role_name
     );
 
-    let response = reqwest::get(&credentials_provider_url);
-
-    match response {
-        Ok(mut resp) => {
-            if resp.status().is_success() {
-                let mut credentials_response = String::new();
-                if resp.read_to_string(&mut credentials_response).is_err() {
-                    return Err(CredentialsError::new(
-                        "Had issues with reading iam role response: {}",
-                    ));
-                }
-
-                parse_credentials_from_aws_service(&credentials_response)
-            } else {
-                return Err(CredentialsError::new(format!(
-                    "Couldn't connect to credentials provider: {}",
-                    resp.status()
-                )));
-            }
-        }
+    let core = try!(Core::new());
+    let uri = try!(credentials_provider_url.parse::<Uri>());
+    match make_request(core, uri, timeout) {
+        Ok(resp) => parse_credentials_from_aws_service(&resp),
         Err(err) => {
             return Err(CredentialsError::new(
                 format!("Couldn't connect to credentials provider: {}", err),