+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Added CHANGELOG
 - Updated CONTRIBUTING to explain PR process
 - Added Application Autoscaling service
+- Added Athena service
+- Added X-Ray service
+- Updated Credentials crate to use hyper 0.11 (aka the Async IO Update).
 
 ## [0.28.0] - 2017-08-25